Identified issue in setting Header or Footer
There is a limitation of around 255 in length of the string that can be
used in header & footer. If it exceeds the limit, then it raises an
exception. Header/Footer must be set accordingly to avoid this
exception.
Modified the font of Header & Footer to Arial to reduce the length.
Fixed issue in Progressbar progress setting.
Added shortcuts to all elements in CreateSellerInvoice form &
AddNewOrderSheet form. Also modified tabindex in both form to be in
sequence.